Refactor/tests (#1342)
Fixing tests and Serf edge case

+++ kong/cmd/utils/serf_signals.lua
@@ -111,8 +111,13 @@ function _M.stop(kong_config, dao)
   local ok, err = serf:leave()
   if not ok then return nil, err end
 
-  log.verbose("stopping Serf agent at %s", kong_config.serf_pid)
-  return kill(kong_config.serf_pid, "-9")
+  if pl_path.exists(kong_config.serf_pid) then
+    log.verbose("stopping Serf agent at %s", kong_config.serf_pid)
+    local code = kill(kong_config.serf_pid, "-9")
+    pl_file.delete(kong_config.serf_pid)
+    return code
+  end
+  return true
 end
 
 return _M
